WinGui: Don't cleanup queue files in multi-instance mode.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk/win/CS@4444 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ Functions/Main.cs
@@ -423,9 +423,12 @@ namespace Handbrake.Functions
                 }
 
                 // Cleanup old/unused queue files for now.
-                foreach (string file in removeFiles)
+                if (!GeneralUtilities.IsMultiInstance)
                 {
-                    File.Delete(file);
+                    foreach (string file in removeFiles)
+                    {
+                        File.Delete(file);
+                    }
                 }
 
                 return queueFiles;
